+++ src/main/java/de/marhali/easyi18n/util/WildcardRegexMatcher.java
@@ -1,0 +1,23 @@
+package de.marhali.easyi18n.util;
+
+import org.apache.commons.io.FilenameUtils;
+
+/**
+ * Utilities for wildcard / regex matching.
+ * @author marhali
+ */
+public class WildcardRegexMatcher {
+    public static boolean matchWildcardRegex(String string, String pattern) {
+        boolean wildcardMatch = FilenameUtils.wildcardMatchOnSystem(string, pattern);
+
+        if(wildcardMatch) {
+            return true;
+        }
+
+        try {
+            return string.matches(pattern);
+        } catch (Exception e) {
+            return false;
+        }
+    }
+}