+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 ### Added
+- Support for IntelliJ Platform version 2022.3
 - Regex support for translation file patterns
 
 ### Changed