+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # easy-i18n Changelog
 
 ## [Unreleased]
+
+## [4.3.0]
 ### Added
 - Support for IntelliJ Platform version 2022.3
 - Regex support for translation file patterns